workaround disabling default 'become' breaking per-action 'become'
In OSCI zone we don't need 'become' at all, so I disabled it.
Unfortunately this also prevented 'become' in action to not work
anymore. It is useful for example when becoming 'postgres' to setup a PG
database.

More investigation is needed, but not disabling make things work (with
the performance penalty of sudoing uselessly).

Change-Id: Ib3ef8c99fe2c0219c001fa1d0b3b4f091976fc22