fix variable clash between postfix and dovecot roles
Ansible loads all role defaults _first_, and then execute the roles.
This is counter-intuitive and results in a role called later to affect
things earlier in the play (as variables are global).

Here more specifically the variables 'auth' is left undefined in the
postfix role and defined to "pam" in the dovecot role. When the postfix
role is executed it thinks an authentication mechanism is to be defined
because the variable holds a value. Due to the unexpected value of the
variable, only the common part of the config is setup, enabling SASL,
but no real mechanism is configured. This result in spammy warnings
when postfix receives mails.

Using include_role allows to load defaults only when the task in run,
having only a forward effect on variables. This is also quite practical
to treat calling roles as a task so they can be mixed and easily
ordered. I found this solution better than artificially splitting the
play for no logical reason.

Change-Id: I2430bda3e45d43f66a23eb3c678e307428ef0aa5