install needrestart
This is used to check for services to restart. yum/dnf-cron is fine to
install security fixes but not all packages properly restart and
non-leaf packages are not handled. Also daemon using interpreted
languages need reverse-dependency detection.

needrestart is not yet activated automatically, it only suggests
restarts after a manual upgrade. I already tested it for years on Debian
and have been testing it for some time in OSAS on Red Hat systems.

Change-Id: Ib39adf17c1d871f6d05c2cbed925b9e4a2b6c483