Add caching to mirrorlist vhost
This patch adds the CacheEnable directive which will
cache responses from the upstream server for 24h (default value)
so even if something happens to resources users will be able
to fetch the mirrorlist. At the moment it is updated statically,
so cached data should be up-to-date.

Note: this requires `cache` and `cache_disk` modules to operate.

Change-Id: Ic708e831006a450250b30fb301d7c814395eda9e
Signed-off-by: Evgheni Dereveanchin <ederevea@redhat.com>